fix config



See merge request !108

+++ mc-core/src/main/java/com/makenv/model/mc/core/bean/MeicParams.java
@@ -49,17 +49,17 @@ public class MeicParams {
 
     private String password;
 
-    private int sleepSeconds;
+    private long sleepSeconds;
 
-    public int getSleepSeconds() {
-        return sleepSeconds;
-    }
+  public long getSleepSeconds() {
+    return sleepSeconds;
+  }
 
-    public void setSleepSeconds(int sleepSeconds) {
-        this.sleepSeconds = sleepSeconds;
-    }
+  public void setSleepSeconds(long sleepSeconds) {
+    this.sleepSeconds = sleepSeconds;
+  }
 
-    public String getUsername() {
+  public String getUsername() {
         return username;
     }
